Bump rimraf to 5.0.5
Older versions of rimraf transitively depended on a package called
inflight, which is no longer maintained, and current has a medium
severity security vulnerability associated with it.

Newer versions of rimraf rely on a newer version of glob, which no
longer imports inflight.

Due to the differences in major version, rimraf has since changed their
API. They now return promises (as opposed to callbacks), and as a
result, we need to provide two functions to the .then() continuation to
invoke the callback correctly (with the parameters in the correct
order).

(Fix originally written by pnappa for the mapbox/node-pre-gyp project. Merged for use in this @discordjs/node-pre-gyp as a tempoary security fix until the main @discordjs/node-pre-gyp project is updated).